Created endpoint to add purchase and purchase details and to return purchases and purchases details

+++ ProductStore/Models/PurchaseDetail.cs
@@ -7,7 +7,7 @@ public partial class PurchaseDetail
 {
     public int Id { get; set; }
 
-    public decimal Number { get; set; }
+    public float Number { get; set; }
 
     public float PriceForOnePiece { get; set; }
